streamlit_app/requirements.txt
streamlit==1.31.0
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.0
streamlit-authenticator==0.2.3
pyyaml==6.0.1
anthropic==0.39.0
twilio==9.0.4 
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0